# swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "paths": {
        "/api/usage": {
            "get": {
                "responses": {
                    "403": {
                        "description": "No permission to access this api"
                    },
                    "404": {
                        "description": "The usage information does not exist"
                    },
                    "200": {
                        "description": "The usage data has generated and returned successfully"
                    }
                },
                "description": "Generate all api usage information",
                "operationId": "get_usage",
                "tags": [
                    "api"
                ]
            }
        },
        "/api/usage/{api_name}": {
            "get": {
                "responses": {
                    "403": {
                        "description": "No permission to access this api"
                    },
                    "404": {
                        "description": "The usage information does not exist"
                    },
                    "200": {
                        "description": "The usage information has return successfully"
                    }
                },
                "description": "Generate specific api usage information",
                "operationId": "get_usage",
                "parameters": [
                    {
                        "name": "api_name",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "description": "The name of the api"
                    }
                ],
                "tags": [
                    "api"
                ]
            }
        },
...
    },
    "info": {
        "title": "API",
        "version": "1.0",
        "description": "Weather prediction service api"
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "securityDefinitions": {
        "API-KEY": {
            "type": "apiKey",
            "in": "header",
            "name": "AUTH-TOKEN"
        }
    },
    "security": [
        {
            "API-KEY": []
        }
    ],
...
}